Agregando campo surname_transmiter a OrderShipment

+++ src/models/TypeShipment.js
@@ -5,7 +5,7 @@ export default (sequelize) => {
     "TypeShipment",
     {
       id: {
-        type: DataTypes.TINYINT,
+        type: DataTypes.SMALLINT,
         primaryKey: true,
         autoIncrement: true,
       },
